requirements.txt
algokit-utils==2.3.1
anyio==4.4.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
py-algorand-sdk==2.6.1
pycparser==2.22
pycryptodomex==3.20.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyNaCl==1.5.0
pytesseract==0.3.13
python-engineio==4.9.1
python-socketio==5.11.4
pyzbar==0.1.9
requests==2.32.3
rfc3986==1.5.0
simple-websocket==1.0.0
sniffio==1.3.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
